+++ modules/canvaskit/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - `CanvasKit.Typeface.GetDefault()` as a way to explicitly get the compiled-in typeface (if any).
+ - `Canvas.quickReject` to quickly check if a Rect is within the current clip region.
 
 ## [0.39.1] - 2023-10-12